Send better stack traces
Prior to this commit, the stack trace sent to Sentry would include references
to all the FireGento Logger code plus this extension's logger. This made it
harder to quickly determine where the actual error/exception occurred.

So now we simply attempt to "rewind" the stack trace to the point where the
error/exception occurred.  And if for any reason we can't, we'll just use the
original functionality.